chore(jwt): refactor jwt service
